requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-docx==1.1.0
